fix(moves): update include paths and tidy includes; fix test warnings
- Use bazel include path `moves/Moves.h` and add //library/src/moves:moves to deps
- Remove unused includes and forward concrete trans_table headers to impl
- Fix test warnings (unused captures/variables)
