+++ library/src/BUILD.bazel
@@ -15,7 +15,8 @@ cc_library(
         "//library/src/utility:constants",
         "//library/src/utility:lookup_tables",
         "//library/src/heuristic_sorting",
-        "//library/src/trans_table",
+    "//library/src/trans_table",
+    "//library/src/moves:moves",
     ],
 )
 
@@ -38,6 +39,7 @@ cc_library(
         "//library/src/utility:constants",
         "//library/src/utility:lookup_tables",
         "//library/src/heuristic_sorting",
-        "//library/src/trans_table",
+    "//library/src/trans_table",
+    "//library/src/moves:moves",
     ],
 )
